feat(app): crypto monetization — Adsterra banners, Coinzilla rewarded video, donation modal
- ad_config.dart: central monetization policy (test-mode master switch,
  zone keys, VAST/direct URLs, grant-on-ad-failure product decision,
  PremiumAction catalogue for the four gated features).
- adsterra_banner_widget.dart: WebView-wrapped Adsterra JS tag in IAB sizes
  (320x50 sticky, 300x250 mid-result, 728x90), navigation delegate keeping
  the frame on the ad host, clearly-labeled test placeholder (no network in
  test mode), StickyAdBar with sponsored chip + session collapse, and a
  MidResultBanner block. Mounted only on non-camera screens.
- coinzilla_rewarded_service.dart: persisted FeatureTokenWallet
  (SharedPreferences + ValueNotifier), namespace-agnostic VAST 2/3/4 parser
  (prefers highest-bitrate MP4, extracts impression/error beacons),
  fullscreen non-dismissable rewarded modal (video_player, 30s minimum watch
  clock, mute toggle, quit confirmation, retry hygiene) with labeled test
  creative (5s), token grant on completion, and ensureEntitlement() used by
  premium actions. Ad-network failure unlocks anyway (documented decision).
- crypto_donation_modal.dart: cyberpunk dialog (scanlines, neon corner
  brackets, glow border) with QR codes (white quiet-zone panels) and 1-tap
  copy + haptics for the maintainer's SOL/BTC/ETH addresses — all three
  validated (BTC bech32 polymod, SOL base58->32-byte ed25519, ETH hex) —
  plus thank-you note from AndrexTheDev.
- monetization_test.dart: wallet grant/spend/persist/overdraft, VAST parsing
  (best-MediaFile selection, garbage rejection), address validators, action
  catalogue.
- main.dart: dashboard sticky banner, wallet chip (earn-a-token flow),
  donation heart in the app bar. pubspec: webview_flutter, video_player,
  qr_flutter, xml.

+++ app/lib/screens/modes/horizon_dip_screen.dart
@@ -440,9 +440,9 @@ class _HorizonDipHudPainter extends CustomPainter {
 
     // Legend.
     _legend(canvas, size, 'green: 0° true horizontal (EKF pitch)',
-        Offset(12, yHorizon0.clamp(20, size.height - 90)));
+        Offset(12, yHorizon0.clamp(20, size.height - 90).toDouble()));
     _legend(canvas, size, 'cyan: globe prediction θ = 1.06′·√h',
-        Offset(12, yGlobe.clamp(20, size.height - 60)));
+        Offset(12, yGlobe.clamp(20, size.height - 60).toDouble()));
 
     if (!hasFix) {
       labelTp.text = const TextSpan(
